refactor(scanner): extract the rescan-worker subsystem into scanner_rescan.go
scanner.go was ~2.3k lines. Move the self-contained pending-rescan subsystem —
queueForRescan, StartRescanWorker, loadPendingRescanFiles, markRescanResolved,
updateRescanRetry, emitRescanCorruption, processPendingRescans,
GetPendingRescanStats, and the pendingRescanFile type — verbatim into
scanner_rescan.go (same package). Pure code move: no logic, signature, or
behavior change. scanner.go drops 2262 -> 2070 lines.

Test plan:
- go build ./... ; golangci-lint ./internal/services/... 0 issues
- go test ./internal/services/ pass
